Restructure map so contents can be passed in, allowing same component to be re-used in ReportForm
